analysis/requirements.txt
anthropic>=0.40.0,<1.0.0
google-genai>=1.0.0,<2.0.0
markdown>=3.5,<4.0.0
matplotlib>=3.5,<4.0.0
numpy>=1.21,<3.0.0
pydantic>=2.0,<3.0.0
pytest>=7.0,<10.0.0
scipy>=1.7.0,<2.0.0
tqdm>=4.60.0,<5.0.0
weasyprint>=60.0,<69.0.0
